requirements.txt
chromadb
python-dotenv
google-generativeai
sentence-transformers
streamlit
pypdf
python-docx
python-pptx
pandas
openpyxl